ml-service/requirements.txt
torch==2.0.0
torchvision==0.15.1
pyyaml
marshmallow-dataclass
grpcio-tools==1.53.0
click